requirements.txt
flask==2.3.3
pdfminer.six==20220524
Werkzeug==2.3.7
gunicorn==21.2.0
flask-session==0.5.0
reportlab==4.0.7  # For test PDF generation
pytest==7.4.0  # For more advanced testing features
coverage==7.3.1  # For test coverage reporting
